feat: complete Phase 4.3 - performance optimization (15 tests passing)
  Sub-phase 4.3: Performance Optimization - COMPLETED
  Phase 4: Key Management and Caching - COMPLETED

  Changes:
  - Fixed compilation error in test_performance.rs (Duration division type cast)
  - Added test_performance module to test runner
  - All 15 performance tests passing (175/175 total EZKL tests)

  Verified implementations:
  - Performance tests (tests/ezkl/test_performance.rs - 514 lines, 15 tests)
  - Metrics implementation (src/crypto/ezkl/metrics.rs - 568 lines, 16 tests)

  Performance test coverage:
  - Proof generation performance (< 1ms actual vs < 10ms target)
  - Witness generation (< 1ms vs < 5ms target)
  - Concurrent proof generation (4+ parallel)
  - Sequential throughput (~1.9M proofs/sec vs > 1000 target)
  - Key loading performance (~12μs vs < 50ms target)
  - Binary serialization (< 100μs)
  - p50/p95/p99 percentile targets met
  - Memory leak detection (1000+ cycles)
  - Performance degradation testing

  Metrics features:
  - Thread-safe atomic counters
  - Proof generation tracking (attempts, success, errors, duration)
  - Cache hit/miss rates (proof cache + key cache)
  - Verification tracking (attempts, success, failures, duration)
  - Prometheus text format export
  - Global singleton with lazy initialization

  Phase 4 Complete: 53 tests (18 key + 20 cache + 15 performance)

+++ tests/ezkl/test_performance.rs
@@ -86,7 +86,7 @@ fn test_batch_proof_generation_performance() -> Result<()> {
     }
 
     let duration = start.elapsed();
-    let avg_time = duration / count;
+    let avg_time = duration / count as u32;
 
     println!("Generated {} proofs in {:?}", count, duration);
     println!("Average time per proof: {:?}", avg_time);
